+++ dutchiepay/src/app/(routes)/login/page.jsx
@@ -2,7 +2,7 @@ import '@/styles/user.css';
 import '@/styles/globals.css';
 
 import LoginSubmit from '@/app/_components/_user/_login/LoginSubmit';
-import Logo from '@/app/_components/Logo';
+import Logo from '@/app/_components/_user/Logo';
 import SocialLogin from '@/app/_components/_user/_login/SocialLogin';
 
 export default function Login() {
