+++ dutchiepay/src/app/(routes)/signup/page.jsx
@@ -2,7 +2,7 @@ import '@/styles/globals.css';
 import '@/styles/user.css';
 
 import Link from 'next/link';
-import Logo from '@/app/_components/Logo';
+import Logo from '@/app/_components/_user/Logo';
 import SignUpSubmit from '@/app/_components/_user/_signup/SignUpSubmit';
 import SocialSignup from '@/app/_components/_user/_signup/SocialSignup';
