+++ dutchiepay/src/styles/user.css
@@ -143,14 +143,14 @@
   text-align: center;
   cursor: pointer;
   transition:
-    border-bottom 0.1s ease,
-    font-weight 0.1s ease;
+    border-bottom 0.2s ease,
+    font-weight 0.2s ease;
 }
 
 .user-find__header__active {
   border-bottom: 2px solid #4b70f5;
   font-weight: bold;
   transition:
-    border-bottom 0.1s ease,
-    font-weight 0.1s ease;
+    border-bottom 0.2s ease,
+    font-weight 0.2s ease;
 }
